+++ conf/config.go
@@ -11,8 +11,8 @@ import (
 )
 
 var (
-	BaseUrl               = "https://testnet-backend.degate.com/"
-	WebsocketBaseUrl      = "wss://testnet-ws.degate.com/"
+	BaseUrl               = "https://mainnet-backend.degate.com"
+	WebsocketBaseUrl      = "wss://mainnet-ws.degate.com"
 	OrderBookPath         = "order-book-api/"
 	WsPath                = "order-book-ws-api/"
 	EffectiveDigits       = 6